Deleted connectivityCheck, minor adjustments visuals notificatino & implemented integration Firestore (in progress)
Connectivitycheck will be replaced by async/await functinos, adjusted speed, delay and padding notification to make it feel nicer and quicker. Finally implemented integratino with firestore. Now posts wille be saved in that database first. BUG:
The function that sends messages to the database does not return correctly; thus these are undefined (it's thus still a work in progress)